HTML `Fetch` function for vpnsecure and slickvpn
