+++ internal/provider/slickvpn/updater/website_test.go
@@ -104,73 +104,6 @@ func Test_fetchServers(t *testing.T) {
 	}
 }
 
-func Test_fetchHTML(t *testing.T) {
-	t.Parallel()
-
-	canceledCtx, cancel := context.WithCancel(context.Background())
-	cancel()
-
-	testCases := map[string]struct {
-		ctx            context.Context
-		responseStatus int
-		responseBody   io.ReadCloser
-		rootNode       *html.Node
-		errWrapped     error
-		errMessage     string
-	}{
-		"context canceled": {
-			ctx:        canceledCtx,
-			errWrapped: context.Canceled,
-			errMessage: `Get "https://www.slickvpn.com/locations/": context canceled`,
-		},
-		"response status not ok": {
-			ctx:            context.Background(),
-			responseStatus: http.StatusNotFound,
-			errWrapped:     ErrHTTPStatusCode,
-			errMessage:     `HTTP status code is not OK: 404 Not Found`,
-		},
-		"success": {
-			ctx:            context.Background(),
-			responseStatus: http.StatusOK,
-			rootNode:       parseTestHTML(t, "some body"),
-			responseBody:   ioutil.NopCloser(strings.NewReader("some body")),
-		},
-	}
-
-	for name, testCase := range testCases {
-		testCase := testCase
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-
-			client := &http.Client{
-				Transport: roundTripFunc(func(r *http.Request) (*http.Response, error) {
-					assert.Equal(t, http.MethodGet, r.Method)
-					assert.Equal(t, r.URL.String(), "https://www.slickvpn.com/locations/")
-
-					ctxErr := r.Context().Err()
-					if ctxErr != nil {
-						return nil, ctxErr
-					}
-
-					return &http.Response{
-						StatusCode: testCase.responseStatus,
-						Status:     http.StatusText(testCase.responseStatus),
-						Body:       testCase.responseBody,
-					}, nil
-				}),
-			}
-
-			rootNode, err := fetchHTML(testCase.ctx, client)
-
-			assert.ErrorIs(t, err, testCase.errWrapped)
-			if testCase.errWrapped != nil {
-				assert.EqualError(t, err, testCase.errMessage)
-			}
-			assert.Equal(t, testCase.rootNode, rootNode)
-		})
-	}
-}
-
 func Test_parseHTML(t *testing.T) {
 	t.Parallel()
